Rename CommandSender class to Protocol to better describe what it is Rename do_discover() to send_handshake() & create a proxy method send_handshake() in Device class
