+++ miio/tests/test_airfresh.py
@@ -10,7 +10,7 @@ from miio.airfresh import (
     OperationMode,
 )
 
-from .dummies import DummyCommandSender, DummyDevice
+from .dummies import DummyDevice, DummyProtocol
 
 
 class DummyAirFresh(DummyDevice, AirFresh):
@@ -36,7 +36,7 @@ class DummyAirFresh(DummyDevice, AirFresh):
             "favorite_level": None,
             "led": "on",
         }
-        self.command_sender = DummyCommandSender(
+        self.protocol = DummyProtocol(
             {
                 "get_prop": self._get_state,
                 "set_power": lambda x: self._set_state("power", x),
