+++ miio/tests/test_philips_rwread.py
@@ -10,7 +10,7 @@ from miio.philips_rwread import (
     PhilipsRwreadStatus,
 )
 
-from .dummies import DummyCommandSender, DummyDevice
+from .dummies import DummyDevice, DummyProtocol
 
 
 class DummyPhilipsRwread(DummyDevice, PhilipsRwread):
@@ -25,7 +25,7 @@ class DummyPhilipsRwread(DummyDevice, PhilipsRwread):
             "flmv": 2,
             "chl": 0,
         }
-        self.command_sender = DummyCommandSender(
+        self.protocol = DummyProtocol(
             {
                 "get_prop": self._get_state,
                 "set_power": lambda x: self._set_state("power", x),
