+++ miio/tests/test_yeelight.py
@@ -5,7 +5,7 @@ import pytest
 from miio import Yeelight
 from miio.yeelight import YeelightException, YeelightMode, YeelightStatus
 
-from .dummies import DummyCommandSender, DummyDevice
+from .dummies import DummyDevice, DummyProtocol
 
 
 class DummyLight(DummyDevice, Yeelight):
@@ -23,7 +23,7 @@ class DummyLight(DummyDevice, Yeelight):
             "save_state": "1",
         }
 
-        self.command_sender = DummyCommandSender(
+        self.protocol = DummyProtocol(
             {
                 "get_prop": self._get_state,
                 "set_power": lambda x: self._set_state("power", x),
