+++ miio/vacuum_cli.py
@@ -19,8 +19,8 @@ from miio.click_common import (
     validate_ip,
     validate_token,
 )
-from miio.command_sender import CommandSender
 from miio.device import UpdateState
+from miio.protocol import Protocol
 from miio.updater import OneShotServer
 
 _LOGGER = logging.getLogger(__name__)
@@ -80,10 +80,10 @@ def cli(ctx, ip: str, token: str, debug: int, id_file: str):
 @cli.resultcallback()
 @pass_dev
 def cleanup(vac: miio.Vacuum, *args, **kwargs):
-    if vac.command_sender.ip is None:  # dummy Device for discovery, skip teardown
+    if vac.protocol.ip is None:  # dummy Device for discovery, skip teardown
         return
     id_file = kwargs["id_file"]
-    seqs = {"seq": vac.command_sender.raw_id, "manual_seq": vac.manual_seqnum}
+    seqs = {"seq": vac.protocol.raw_id, "manual_seq": vac.manual_seqnum}
     _LOGGER.debug("Writing %s to %s", seqs, id_file)
     path_obj = pathlib.Path(id_file)
     dir = path_obj.parents[0]
@@ -100,7 +100,7 @@ def cleanup(vac: miio.Vacuum, *args, **kwargs):
 def discover(handshake):
     """Search for robots in the network."""
     if handshake:
-        CommandSender.discover()
+        Protocol.discover()
     else:
         miio.Discovery.discover_mdns()
